feat: omg! add queue processing fargate service
adds the pickup worker stack that creates a queue-processing-fargate-service.

The idea is to recreate the gist of what the aws-copilot cli would have created for a "worker service" pattern, but in aws-cdk code so it can be deployed with SST and we can share constructs between the lambda + dynamo api and the pickup+ipfs workers.

- https://github.com/ipfs-shipyard/go-ipfs-docker-examples/blob/main/gateway-copilot-backend-service/copilot/gw/manifest.yml
- https://github.com/olizilla/pickup/pull/1
- https://docs.aws.amazon.com/cdk/v2/guide/ecs_example.html
- https://github.com/aws/aws-cdk/blob/main/packages/@aws-cdk/aws-ecs-patterns/lib/fargate/queue-processing-fargate-service.ts

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>