feat: write jobs to sqs queue
IT WORKS ✨🎷🐩

when linked up to the https://github.com/olizilla/pickup this PR has succesfully turned a pin request api call into a dynamoDB record and an SQS Queue message. Pickup pulled the message, fetched the cid via the local go-ipfs sidecar and uploaded the CAR to s3.

License: (Apache-2.0 AND MIT)
Signed-off-by: Oli Evans <oli@tableflip.io>